refactor: move source files to src directory
