--- mention.js
@@ -1,65 +1,0 @@
-import dotenv from 'dotenv';
-import getAnswer from './answer.js';
-import {
-  addPrediction,
-  getPredictions,
-  updatePrediction,
-  deletePrediction,
-} from './db.js';
-dotenv.config();
-
-const text = process.argv.slice(2).join(' ').trim();
-if (!text) {
-  console.error('Usage: npm run mention "<message>"');
-  process.exit(1);
-}
-
-const USER_ID = 'local-user';
-
-async function main() {
-  const match = /agi will arrive on (.+)/i.exec(text);
-  if (match) {
-    const when = match[1].trim();
-    addPrediction(USER_ID, when);
-    console.log(`<@${USER_ID}> Prediction saved for ${when}`);
-    return;
-  }
-
-  const updateMatch = /^update prediction (\d+) to (.+)/i.exec(text);
-  if (updateMatch) {
-    const id = parseInt(updateMatch[1], 10);
-    const when = updateMatch[2].trim();
-    updatePrediction(id, when);
-    console.log(`<@${USER_ID}> Prediction ${id} updated to ${when}`);
-    return;
-  }
-
-  const deleteMatch = /^delete prediction (\d+)/i.exec(text);
-  if (deleteMatch) {
-    const id = parseInt(deleteMatch[1], 10);
-    deletePrediction(id);
-    console.log(`<@${USER_ID}> Prediction ${id} deleted`);
-    return;
-  }
-
-  if (/^list agi predictions$/i.test(text)) {
-    const predictions = getPredictions();
-    if (predictions.length === 0) {
-      console.log('No AGI predictions saved.');
-    } else {
-      const lines = predictions
-        .map((p) => `<@${p.userId}> predicted ${p.date}`)
-        .join('\n');
-      console.log(lines);
-    }
-    return;
-  }
-
-  const answer = await getAnswer(text);
-  console.log(`<@${USER_ID}> ${answer}`);
-}
-
-main().catch((err) => {
-  console.error('Error:', err);
-  process.exit(1);
-});
