+++ package.json
@@ -5,12 +5,12 @@
   "license": "ISC",
   "author": "Enrique Moreno Tent",
   "type": "module",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
-    "start": "node index.js",
-    "hf-test": "node test-fetch.js",
-    "mention": "node mention.js",
+    "start": "node src/index.js",
+    "hf-test": "node src/test-fetch.js",
+    "mention": "node src/mention.js",
     "lint": "eslint ."
   },
   "dependencies": {
