+++ src/index.js
@@ -1,0 +1,4 @@
+// Entry point that starts the Discord bot
+import { startBot } from './bot.js';
+
+startBot();
